Modify the title, author, and pages of book card
Rearrange the label and detail such as title of the book
for ease of reading the book card. It used to be
"Title: {book title}" all in one line, now it is
separated into two lines -- below the book title
is the Title label. The same modification goes
for author and pages.