Edit the layout, size, and color of book card
Change the layout to grid for ease of arranging the
book details such as the title, author, pages, etc.
The title, author, and pages are larger compared to
the buttons to give room for a longer title and author
names. Also changes the weight and style of the labels
to easily distinguish them. The color of buttons when
being hovered is also changed to a darker versio of their
respective colors.